+++ oar/kao/meta_sched.py
@@ -259,7 +259,6 @@ def prepare_job_to_be_launched(session, config, job, current_time_sec):
 def handle_waiting_reservation_jobs(
     session, config, queue_name, resource_set, job_security_time, current_time_sec
 ):
-
     reservation_waiting_timeout = int(config["RESERVATION_WAITING_RESOURCES_TIMEOUT"])
     logger.debug(
         "Queue " + queue_name + ": begin processing accepted Advance Reservations"
@@ -407,7 +406,7 @@ def check_reservation_jobs(
             slots_set = all_slot_sets[ss_name]
 
             t_e = job.start_time + walltime - job_security_time
-            sid_left, sid_right = slots_set.get_encompassing_range(job.start_time, t_e)
+            sid_left, sid_right = slots_set.get_encompassing_slots(job.start_time, t_e)
 
             slots = slots_set.slots
 
@@ -939,8 +938,8 @@ def meta_schedule(session, config, mode="internal", plt=Platform()):
         )
 
         logger.debug(
-            "Queue(s): {},  Launching scheduler, at time: {} ".format(
-                " ".join([q.name for q in queues]), initial_time_sql
+            "Queue(s): {},  Launching scheduler, at time: {}, ({})".format(
+                " ".join([q.name for q in queues]), initial_time_sql, initial_time_sec
             )
         )
 
@@ -1029,10 +1028,18 @@ def meta_schedule(session, config, mode="internal", plt=Platform()):
 
     # Filter jobs that are not yet ready to be scheduled, but present because of the
     # kill_duration_before_reservation=kill_duration_before_reservation parameter
-    jobs_to_launch_lst = filter(
-        lambda j: j.start_time <= current_time_sec,
-        jobs_to_launch_with_security_time_lst,
-    )
+    # filter on date and test that all resource is not absent (alive)
+    jobs_to_launch_lst = [
+        job
+        for job in filter(
+            lambda j: (j.start_time <= current_time_sec)
+            and ((resource_set.absent_roid_itvs & j.res_set) == ProcSet()),
+            jobs_to_launch_with_security_time_lst,
+        )
+    ]
+
+    # for job in jobs_to_launch_lst:
+    #    logger.debug(f"TOLAUNCH Job id:{job.id}, start_time: {job.start_time}, now: {current_time_sec}")
 
     if (
         check_besteffort_jobs_to_kill(