+++ oar/kao/scheduling_basic.py
@@ -45,7 +45,7 @@ def find_first_suitable_contiguous_slots(
     else:
         sid_left = 0
 
-    for (slot_begin, slot_end) in slots_set.traverse_with_width(walltime):
+    for slot_begin, slot_end in slots_set.traverse_with_width(walltime):
         # find next contiguous slots_time
         itvs_avail = intersec_itvs_slots(slots, slot_begin.id, slot_end.id)
         itvs = find_resource_hierarchies_job(itvs_avail, hy_res_rqts, hy)
@@ -95,7 +95,7 @@ def assign_resources_mld_job_split_slots(slots_set: SlotSet, job, hy):
             prev_t_finish = t_finish
             prev_res_set = res_set
             prev_res_rqt = res_rqt
-            (prev_sid_left, prev_sid_right) = slots_set.get_encompassing_range(
+            (prev_sid_left, prev_sid_right) = slots_set.get_encompassing_slots(
                 prev_start_time, prev_t_finish
             )