+++ oar/modules/greta.py
@@ -41,6 +41,7 @@ import pickle
 import re
 import socket
 import sys
+import traceback
 from multiprocessing import Pool, TimeoutError
 from typing import List, Union
 
@@ -457,6 +458,7 @@ class Greta(object):
                 if cmd == "WAKEUP":
                     # Save the timeout for the nodes to be processed.
                     cmd_info["timeout"] = tools.get_date(session) + timeout
+                    logger.debug(f"wakeup timeout: '{timeout}'")
                     command_toLaunch.append(("WAKEUP", node))
                 elif cmd == "HALT":
                     # Don't halt nodes that needs to be kept alive
@@ -506,7 +508,9 @@ class Greta(object):
             # (disable HALT cmd to satisfy keepAlive condition)
 
             for node in nodes_toRemove_from_list_to_process:
-                del nodes_list_to_process[node]
+                n = nodes_list_to_process.get(node, None)
+                if n:
+                    del nodes_list_to_process[node]
 
             # Launching commands
             if command_toLaunch:
@@ -533,9 +537,11 @@ class Greta(object):
                         "nodes_list_to_remind": nodes_list_to_remind,
                     }
                     pickle.dump(greta_status_dump, dump_file, pickle.HIGHEST_PROTOCOL)
+                logger.debug("Max cycles reached: suicide. bye bye.")
                 return 42
 
             if not loop:
+                logger.debug("Exiting main loop normally.")
                 break
         return 0
 
@@ -667,7 +673,13 @@ def main():  # pragma: no cover
     if greta.exit_code:
         return greta.exit_code
 
-    return greta.run()
+    try:
+        status = greta.run()
+    except Exception as err:
+        logger.error(f"Unexpected {err=}, {type(err)=}; {traceback.format_exc()}")
+        raise
+
+    return status
 
 
 if __name__ == "__main__":  # pragma: no cover